added enable function
I added a pin header to short the TPS5410's ENABLE pin to GND, thus switching off the DC/DC converter, effectively turning the controller completely off.
The controller should function normally, when this is left open. Including a switch that connects those two pins is a way of switching the controller off without disconnecting the 12V supply. Whatever it'll be good for.