added six pin connector for programming
changed the three pin programming header to a six pin version, allowing to directly plug in one of the universal FTDI based USB to UART adaptors.
Pinout is:
1: GND
2: NC
3: NC
4: RX
5: TX
6: NC

with RX and TX being from the ESP's point of view and Pin1 being on the far side from the ESP.

This is currently untested. There was some moving around of components on the DC/DC converter but I don't think that made for any substantial change