added VCC and G15 to pin header
in the earlier design, I had omitted a VCC pin at the pin header. That could be a useful pin for folks using the board with the TAMSOTA firmware, opening up a few pins for external sensors.
Now G0, G2, G15 and G16 are available on the header.
When using them, you should be aware that G0, G2 and G16 are internally pulled up and G15 is pulled low. depending on the loaded firmware, changing the level of those pins (especially at boot time) may trigger unintended behaviour.